backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.52
botocore==1.35.52
celery==5.4.0
click==8.1.7
distro-info==1.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==33.1.0
Flask==3.0.3
importlib_metadata==8.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.25.1
rpds-py==0.21.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.0.6
zipp==3.20.2